refactor(websocket-utils): share createWebSocketRpcChannel
Add a single createWebSocketRpcChannel to @typeagent/websocket-utils/rpcChannel
(built on agent-rpc createChannelAdapter). The studio-service server and the
studio extension client both re-export it, deleting two identical hand-rolled
WebSocket-to-RpcChannel adapters.
