+++ ts/packages/studio-service/src/studioServiceServer.ts
@@ -5,8 +5,8 @@ import { WebSocketServer, WebSocket } from "ws";
 import { AddressInfo } from "node:net";
 import registerDebug from "debug";
 import { createRpc } from "@typeagent/agent-rpc/rpc";
-import type { RpcChannel } from "@typeagent/agent-rpc/channel";
 import { createAgentOriginAllowlist } from "@typeagent/websocket-utils/originAllowlist";
+import { createWebSocketRpcChannel } from "@typeagent/websocket-utils/rpcChannel";
 import type {
     StudioRuntime,
     StudioClientCallFunctions,
@@ -222,49 +222,4 @@ function parseBearerToken(
 }
 
 /** Adapt a `ws` WebSocket to the `agent-rpc` {@link RpcChannel} interface. */
-export function createWebSocketRpcChannel(socket: WebSocket): RpcChannel {
-    const messageHandlers = new Set<(message: any) => void>();
-    const disconnectHandlers = new Set<() => void>();
-    const onceMessage = new Set<(message: any) => void>();
-    const onceDisconnect = new Set<() => void>();
-
-    socket.on("message", (data) => {
-        let message: unknown;
-        try {
-            message = JSON.parse(data.toString());
-        } catch {
-            return; // ignore non-JSON frames
-        }
-        for (const h of messageHandlers) h(message);
-        for (const h of onceMessage.values()) {
-            onceMessage.delete(h);
-            h(message);
-        }
-    });
-    socket.on("close", () => {
-        for (const h of disconnectHandlers) h();
-        for (const h of onceDisconnect.values()) {
-            onceDisconnect.delete(h);
-            h();
-        }
-    });
-
-    return {
-        on(event: "message" | "disconnect", cb: any) {
-            (event === "message" ? messageHandlers : disconnectHandlers).add(
-                cb,
-            );
-        },
-        once(event: "message" | "disconnect", cb: any) {
-            (event === "message" ? onceMessage : onceDisconnect).add(cb);
-        },
-        off(event: "message" | "disconnect", cb: any) {
-            (event === "message" ? messageHandlers : disconnectHandlers).delete(
-                cb,
-            );
-        },
-        send(message: unknown, cb?: (err: Error | null) => void) {
-            socket.send(JSON.stringify(message), (err) => cb?.(err ?? null));
-        },
-    };
-}
+export { createWebSocketRpcChannel };
